+++ src/ingestor.hpp
@@ -61,9 +61,10 @@ namespace WebStat {
 		virtual ~Ingestor();
 		SPECIAL_MEMBERS_DELETE(Ingestor);
 
-		using ScanResult = decltype(scn::scan<std::string_view, std::string_view, uint64_t, std::string_view,
-				QuotedString, QueryString, std::string_view, unsigned short, unsigned int, unsigned int, CLFString,
-				CLFString, CLFString>(std::declval<std::string_view>(), ""));
+		using ScanResult
+				= decltype(scn::scan<std::string_view, std::string_view, uint64_t, std::string_view, QuotedString,
+						QueryString, std::string_view, uint16_t, uint64_t, uint32_t, CLFString, CLFString, CLFString>(
+						std::declval<std::string_view>(), ""));
 		using ScanValues = std::remove_cvref_t<decltype(std::declval<WebStat::Ingestor::ScanResult>()->values())>;
 
 		[[nodiscard]] static ScanResult scanLogLine(std::string_view);
